requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
et_xmlfile==2.0.0
iniconfig==2.3.0
openpyxl==3.1.5
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
